# cloud/openapi/shakecloud.yaml
openapi: 3.1.0
...
info:
  title: shake-cloud API
  version: 0.1.0
  description: |
    Login, access keys, the audit log and compute. Instances can be created
    with free-form cpu, memory and disk; named instance types are only a
    shorthand starting point, not a fixed catalogue of sizes. Quotas and
    capacity limits are administrator-settable at runtime through
    /v1/limits, not baked into the deployment.

    Browsers log in through Authentik and hold a session cookie. Terraform
    and the CLI send `Authorization: Bearer sca_<access_key_id>.<secret>`.
    Keys work while Authentik is down.
servers:
  - url: http://192.168.10.205:8080
security:
  - accessKey: []
  - session: []
...
paths:
...
  /v1/images:
    get:
      operationId: DescribeImages
      tags: [compute]
      summary: List the deployment's shared images and every account's uploaded ones.
      description: 'So people can see what is available to launch from, and who added it.'
      responses:
        '200':
          description: Images.
          content:
            application/json:
              schema:
                type: object
                required: [images]
                properties:
                  images:
                    type: array
                    items: {$ref: '#/components/schemas/Image'}
        '401': {$ref: '#/components/responses/Error'}
    post:
      operationId: ImportImage
      tags: [compute]
      summary: Upload a disk image.
      description: |
        The image is streamed straight through to the node's image store,
        never buffered in the API, because the API container has only a
        small RAM-backed /tmp. Proxmox validates the file with qemu-img, so
        a corrupt or unsupported file is refused and nothing is stored.
        qcow2 and raw are what the store accepts.

        The request is answered when the store has accepted the image, so a
        large upload holds the connection open for minutes.

        The size is bounded by the administrator-settable
        `capacity.max_image_gib` limit and by the free space the image store
        must keep (`capacity.image_store_min_free_mib`).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [name, file]
              properties:
                name: {type: string, maxLength: 128}
                file: {type: string, format: binary}
            encoding:
              file:
                contentType: application/octet-stream
      responses:
        '201':
          description: Uploaded.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ImageEnvelope'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '413': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
  /v1/images/{image_id}:
    delete:
      operationId: DeleteImage
      tags: [compute]
      summary: Delete an uploaded image.
      description: |
        An owner or a cloud-admin may delete an uploaded image; the
        deployment's **shared** images cannot be deleted through the API
        (they are declared in Terraform). Instances already created from an
        image are unaffected, because their disk was copied at creation. A
        deletion is refused while an instance that was launched from it is
        still being created.
      parameters:
        - $ref: '#/components/parameters/ImageID'
      responses:
        '204':
          description: Deleted.
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
...
components:
  parameters:
...
    ImageID:
      name: image_id
      in: path
      required: true
      schema: {type: string, pattern: '^img-[a-z0-9-]+$'}
...
  securitySchemes:
    accessKey:
      type: http
      scheme: bearer
      bearerFormat: sca_<access_key_id>.<secret>
    session:
      type: apiKey
      in: cookie
      name: shakecloud_session
  responses:
    Error:
      description: |
        Codes: AuthFailure (401), UnauthorizedOperation (403),
        CrossOriginRequestBlocked (403), NoSuchEntity (404), ValidationError,
        InvalidParameterValue and MalformedRequest (400), IncorrectInstanceState
        and LimitExceeded (409), InternalError (500), ServiceUnavailable (503).
        Also InvalidImageID.NotFound (404), ImageLimitExceeded (409),
        RequestEntityTooLarge (413), InvalidKeyPair.Duplicate (409) and
        InvalidKeyPair.NotFound (404).
        Volumes and security groups add InvalidVolume.NotFound (404),
        IncorrectState (409), VolumeInUse (409), VolumeLimitExceeded (409),
        AttachmentLimitExceeded (409), InvalidInstanceID.NotFound (404),
        IncorrectInstanceState (409), IdempotentParameterMismatch (409),
        InvalidGroup.NotFound (404), InvalidGroup.Duplicate (409),
        InvalidPermission.Duplicate (409), InvalidPermission.NotFound (404),
        DependencyViolation (409), CannotDelete (409),
        InsufficientVolumeCapacity (503),
        RulesPerSecurityGroupLimitExceeded (409),
        SecurityGroupsPerInstanceLimitExceeded (409) and
        SecurityGroupLimitExceeded (409).
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
  schemas:
    Error:
      type: object
      required: [error, request_id]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
        request_id: {type: string}
...
    Image:
      type: object
      required: [image_id, name, state, public]
      properties:
        image_id: {type: string, pattern: '^img-[a-z0-9-]+$'}
        name: {type: string}
        state: {type: string, enum: [available]}
        public: {type: boolean, description: 'True for the deployment''s shared images, which every account may launch from and nobody may delete through the API.'}
        account_id: {type: string, description: 'The uploader; absent for shared images.'}
        owner_username: {type: string}
        format: {type: string, description: 'qcow2 or raw, as reported by the store.'}
        size_mib: {type: integer}
        created_at: {type: string, format: date-time}
    ImageEnvelope:
      type: object
      required: [image]
      properties:
        image: {$ref: '#/components/schemas/Image'}
...